Add files via upload
Has trouble with longer words where you will have all the letters guessed but it wont show the "congrats" message. Happens when you have some incorrect guesses mixed in with the correct ones. Also doesn't like spaces. Would be cool if the app could be linked to dictionary.com or something and a random word be chosen from there, or any library of words.